refactored test: change fs.defaultFs in test for any config.
